requirements.txt
firecrawl-py
beautifulsoup4
pandas
python-dotenv
lxml
google-cloud-storage
